Extract OpenEO service and validation utilities
Restructure monolithic `code-runner.ts` by moving OpenEO API interactions, service lifecycle management, and process graph validation into dedicated modules under `app/utils/openeo/`.
